+++ open/refine/refine.py
@@ -26,8 +26,8 @@ import time
 import urllib.request, urllib.parse, urllib.error, urllib.response
 import requests
 
-from google.refine import facet
-from google.refine import history
+from open.refine import facet
+from open.refine import history
 
 REFINE_HOST = os.environ.get('OPENREFINE_HOST', os.environ.get('GOOGLE_REFINE_HOST', '127.0.0.1'))
 REFINE_PORT = os.environ.get('OPENREFINE_PORT', os.environ.get('GOOGLE_REFINE_PORT', '3333'))
